+++ build.sh
@@ -17,6 +17,7 @@ swiftc \
   -O \
   -framework AppKit \
   -framework CoreGraphics \
+  -framework CoreText \
   -framework PDFKit \
   -framework SwiftUI \
   -framework UniformTypeIdentifiers \